Add initial setup for frontend
Add new folders in our project-src folder to help organize our frontend. This includes a new folder to hold each page, and a new folder to hold components like the header/footer. Added the files for the different pages. Uploaded an image to the 'assets' folder that contains the logo for our website. Set up the App.jsx file to hold each page of our website, as well as our header/footer.