Update file structure
Changed the file structure to help organize the code for better reusability/modifiability. Removed some dead code (for example, importing react on pages where it was not necessary). Removed the “preview” page that just contained a view of some of the reusable UI elements. Added the title and background for the homepage in the  HomePage.jsx file.