Add file upload component
Added a component where the users can drag and drop files called FileUploadArea.jsx in the components folder. Also includes a button where users can manually select the file they want to choose. Put this component on the HomePage.jsx page and formatted it so that it would fit with the text box component.