Add visual select page
Created a new folder in the components folder called 'visualselect'. Created two files in this folder: VisualSelectStepper.jsx which has the stepper for the page, and ChartSelectionCard.jsc which is the card that displays the chart options, as well as a title and short description.

Updated the VisualSelect.jsx page to include the basic layout. Since the chart API has not been implemented yet, a placeholder is being used. Includes logic for how to connect with the backend, but is not currently functional with the backend. A hardcoded list of sample info is being used as 'fetch' is not set up to work.