Add error message component
Add a new component DefaultError that will be the standard style for our errors. Added a hook for this component called DefaultErrorHook.js  so it can be easily reused. Added the error message in HomePage.jsx file, this specific error will occur when a user tries to view their data when they have not yet uploaded anything.