Add logic to validate files
Updated FileUploadArea.jsx to correctly show users the types of files theycan upload. Also changed the logic so that only valid files will be accepted. If an invalid file type is given, an error message will be displayed using the DefaultError.jsx component.

Updated HomePage.jsx to pass the error message to the upload area component so that the message can be displayed on the main page.