Update uploading logic
Modified HomePage.jsx, FileUploadArea.jsx, and TextInputArea.jsx. Made changes so that users can only upload a file or input text, not both. Set it up so that once they do one, the other gets whited out. If the user still tried to upload, it gives an error message explaining the problem.